+++ sye-azure/lib/machine.ts
@@ -33,6 +33,7 @@ import { VirtualMachine } from 'azure-arm-compute/lib/models'
 import { exit, syeEnvironmentFile } from '../../lib/common'
 
 export async function machineAdd(
+    profile: string,
     clusterId: string,
     region: string,
     availabilityZone: string,
@@ -52,7 +53,7 @@ export async function machineAdd(
 
     validateClusterId(clusterId)
 
-    let credentials = await getCredentials(clusterId)
+    let credentials = await getCredentials(profile)
 
     const subscription = await getSubscription(credentials, { resourceGroup: clusterId })
 
@@ -236,10 +237,10 @@ ROLES="${roles}" PUBLIC_STORAGE_URL="${publicStorageUrl}" SYE_ENV_URL="${envUrl}
     }
 }
 
-export async function machineDelete(clusterId: string, machineName: string, skipSecurityRules = false) {
+export async function machineDelete(profile: string, clusterId: string, machineName: string, skipSecurityRules = false) {
     validateClusterId(clusterId)
 
-    let credentials = await getCredentials(clusterId)
+    let credentials = await getCredentials(profile)
     const subscription = await getSubscription(credentials, { resourceGroup: clusterId })
 
     const networkClient = new NetworkManagementClient(credentials, subscription.subscriptionId)
@@ -283,7 +284,9 @@ export async function machineDelete(clusterId: string, machineName: string, skip
     }
 }
 
-export async function machineRedeploy(_clusterId: string, _region: string, _name: string) {}
+export async function machineRedeploy(_profile: string, _clusterId: string, _region: string, _name: string) {
+    throw new Error('Not yet implemented!')
+}
 
 export async function ensureMachineSecurityRules(clusterId: string) {
     validateClusterId(clusterId)